A VERY IMPORTANT CHANGE TOWARDS MODERNIZATION!
The code uses Fortran's MPI module instead of the
obsolete statement 'include mpif.h'.  The latter is
actually for FORTRAN 77, doesn't do any checks and
its usage is discouraged.

Upon using Fortran's MPI module (use Mpi), compiler
has discovered that quite a number of long integers
(8 bytes) were passed to MPI functions where they
shouldn't have.  In order to work around it, I removed
all implicit switching to long integers from makefiles,
making them all short (4 bytes) and therefore compatible
with MPI functions.  The only variable which should be
8 bytes long is file displacement, so I turned only that
to long integer declaring it with 'integer(DP) :: disp'

In addition to that, I replaced all occurences of
functions operating on shorts (such as Swap_Short,
Sort_Short_Array and alike) and replaced all occurences
of MPI_INTEGER8 to MPI_INTEGER.  Also, MPI_LOGICAL8
is discarded, which is really good because the code
should now be compatible with MPICH.

I have also moved defintions of 'IP' and 'RP' (for
integer precision and real precision) to one place in
Const_Mod.

Eventually, I also removed six MPI functions which were
not used at all in the program.

At the point of this commit, I am only sure it compiles
I didn't run a single case with it.

On branch bojan_uses_mpi_mod_integers_single

modified:   Sources/Convert/makefile
modified:   Sources/Divide/makefile
modified:   Sources/Generate/makefile
modified:   Sources/Process/Backup_Mod/Load.f90
modified:   Sources/Process/Backup_Mod/Read_Cell_Real.f90
modified:   Sources/Process/Backup_Mod/Read_Int.f90
modified:   Sources/Process/Backup_Mod/Read_Int_Array.f90
modified:   Sources/Process/Backup_Mod/Read_Log.f90
modified:   Sources/Process/Backup_Mod/Read_Log_Array.f90
modified:   Sources/Process/Backup_Mod/Read_Real.f90
modified:   Sources/Process/Backup_Mod/Read_Real_Array.f90
modified:   Sources/Process/Backup_Mod/Read_Swarm.f90
modified:   Sources/Process/Backup_Mod/Read_Variable.f90
modified:   Sources/Process/Backup_Mod/Save.f90
modified:   Sources/Process/Backup_Mod/Write_Cell_Real.f90
modified:   Sources/Process/Backup_Mod/Write_Face_Real.f90
modified:   Sources/Process/Backup_Mod/Write_Int.f90
modified:   Sources/Process/Backup_Mod/Write_Int_Array.f90
modified:   Sources/Process/Backup_Mod/Write_Log.f90
modified:   Sources/Process/Backup_Mod/Write_Log_Array.f90
modified:   Sources/Process/Backup_Mod/Write_Real.f90
modified:   Sources/Process/Backup_Mod/Write_Real_Array.f90
modified:   Sources/Process/Backup_Mod/Write_Swarm.f90
modified:   Sources/Process/Backup_Mod/Write_Variable.f90
modified:   Sources/Process/Results_Mod/Save_Results.f90
modified:   Sources/Process/Results_Mod/Save_Scalar_Int.f90
modified:   Sources/Process/Results_Mod/Save_Scalar_Real.f90
modified:   Sources/Process/Results_Mod/Save_Vector_Real.f90
modified:   Sources/Process/makefile
modified:   Sources/Shared/Comm_Mod/Parallel/Exchange_Int_Array.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Global_Max_Int.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Global_Min_Int.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Global_Sum_Int.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Global_Sum_Int_Array.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Read_Bnd_Real.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Read_Cell_Real.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Read_Face_Real.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Read_Int.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Read_Int_Array.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Read_Log.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Read_Log_Array.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Read_Real.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Read_Real_Array.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Read_Text.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Recv_Int_Array.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Send_Int_Array.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Sendrecv_Int_Arrays.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Write_Bnd_Real.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Write_Cell_Real.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Write_Face_Real.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Write_Int.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Write_Int_Array.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Write_Log.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Write_Log_Array.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Write_Real.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Write_Real_Array.f90
modified:   Sources/Shared/Comm_Mod/Parallel/Write_Text.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Read_Bnd_Real.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Read_Cell_Real.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Read_Face_Real.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Read_Int.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Read_Int_Array.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Read_Log.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Read_Log_Array.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Read_Real.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Read_Real_Array.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Read_Text.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Write_Bnd_Real.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Write_Cell_Real.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Write_Face_Real.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Write_Int.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Write_Int_Array.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Write_Log.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Write_Log_Array.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Write_Real.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Write_Real_Array.f90
modified:   Sources/Shared/Comm_Mod/Sequential/Write_Text.f90
modified:   Sources/Shared/Comm_Mod_Par.f90
modified:   Sources/Shared/Comm_Mod_Seq.f90
modified:   Sources/Shared/Const_Mod.f90
modified:   Sources/Shared/Grid_Mod/Save_Debug_Vtu.f90
modified:   Sources/Shared/Grid_Mod/Save_Vtu_Cells.f90
modified:   Sources/Shared/Grid_Mod/Save_Vtu_Faces.f90
modified:   Sources/Shared/Sort_Mod.f90
deleted:    Sources/Shared/Sort_Mod/Short_Array.f90
deleted:    Sources/Shared/Sort_Mod/Short_Carry_Short.f90
deleted:    Sources/Shared/Swap_Short.f90
modified:   Sources/Utilities/Save_Poly_Vtu_Binary.f90

+++ Sources/Shared/Comm_Mod/Parallel/Read_Bnd_Real.f90
@@ -8,12 +8,12 @@
   class(Comm_Type) :: Comm
   integer          :: fh         ! file handle
   real             :: array(:)
-  integer          :: disp       ! displacement in bytes
+  integer(DP)      :: disp       ! displacement in bytes
 !-----------------------------------[Locals]-----------------------------------!
   integer :: error = 0
 !==============================================================================!
 
-  ! Set view for distributed boundary cell data 
+  ! Set view for distributed boundary cell data
   ! (this part is the same as in Write counterpart)
   call Mpi_File_Set_View(fh,                        &
                          disp,                      &
