+++ Sources/Shared/Comm_Mod/Parallel/Write_Text.f90
@@ -8,7 +8,7 @@
   class(Comm_Type) :: Comm
   integer          :: fh            ! file handle
   character        :: text_out*(*)  ! text to write out
-  integer          :: disp          ! displacement in bytes
+  integer(DP)      :: disp          ! displacement in bytes
 !-----------------------------------[Locals]-----------------------------------!
   integer :: leng
   integer :: error = 0
@@ -17,11 +17,11 @@
   leng = len(text_out)
 
   ! Set it at position disp (same as in Read counterpart)
-  call Mpi_File_Set_View(fh,             &   
-                         disp,           &   
+  call Mpi_File_Set_View(fh,             &
+                         disp,           &
                          MPI_CHARACTER,  &
-                         MPI_CHARACTER,  &   
-                         'native',       &   
+                         MPI_CHARACTER,  &
+                         'native',       &
                          MPI_INFO_NULL,  &
                          error)
 
@@ -29,9 +29,9 @@
   call Mpi_File_Write(fh,                 &
                       text_out,           &
                       leng,               &
-                      MPI_CHARACTER,      &   
+                      MPI_CHARACTER,      &
                       MPI_STATUS_IGNORE,  &
-                      error) 
+                      error)
 
   disp = disp + leng
