+++ Sources/Shared/Comm_Mod/Sequential/Read_Int.f90
@@ -8,7 +8,7 @@
   class(Comm_Type) :: Comm
   integer          :: fh    ! file handle
   integer          :: num   ! number to read
-  integer          :: disp  ! displacement in bytes
+  integer(DP)      :: disp  ! displacement in bytes
 !==============================================================================!
 
   ! Position yourself at the right place inside the file
